move W_Exception & co into their own file
