+++ spy/backend/spy.py
@@ -6,7 +6,7 @@ from spy.vm.vm import SPyVM
 from spy.vm.object import W_Object, W_Type
 from spy.vm.function import W_ASTFunc, FuncParam
 from spy.vm.list import W_List
-from spy.vm.modules.builtins import W_Exception
+from spy.vm.exc import W_Exception
 from spy.irgen.scope import SymTable
 from spy.util import magic_dispatch
 from spy.textbuilder import TextBuilder
